reorganized code of token blacklisting
